Big commit with some stuff in it

+++ Assets/Scripts/Resources/Resource_Input.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 [System.Serializable]
 public class Resource_Input{
 	public int inputID;
-	public int amount;
+	public double amount;
 }
